+++ src/Request/ServerRequest.php
@@ -27,16 +27,18 @@ declare(strict_types=1);
 
 namespace Marmotte\Http\Request;
 
+use Marmotte\Brick\Services\Service;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\StreamInterface;
 use Psr\Http\Message\UploadedFileInterface;
 use Psr\Http\Message\UriInterface;
 
+#[Service(autoload: false)]
 final class ServerRequest extends Request implements ServerRequestInterface
 {
     /**
-     * @param UploadedFileInterface[] $uploaded_files
-     * @param array<string, string[]> $headers
+     * @param array<string, UploadedFileInterface> $uploaded_files
+     * @param array<string, string[]>              $headers
      */
     public function __construct(
         private readonly array    $server_params,
@@ -87,7 +89,7 @@ final class ServerRequest extends Request implements ServerRequestInterface
     }
 
     /**
-     * @return UploadedFileInterface[]
+     * @return array<string, UploadedFileInterface>
      */
     public function getUploadedFiles(): array
     {
@@ -97,7 +99,7 @@ final class ServerRequest extends Request implements ServerRequestInterface
     /**
      * @psalm-suppress MoreSpecificImplementedParamType,LessSpecificImplementedReturnType
      *
-     * @param UploadedFileInterface[] $uploadedFiles
+     * @param array<string, UploadedFileInterface> $uploadedFiles
      */
     public function withUploadedFiles(array $uploadedFiles): ServerRequestInterface
     {