+++ src/Response/ResponseFactory.php
@@ -27,12 +27,18 @@ declare(strict_types=1);
 
 namespace Marmotte\Http\Response;
 
+use Marmotte\Brick\Services\Service;
 use Marmotte\Http\Stream\StreamFactory;
 use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ResponseInterface;
 
+#[Service]
 final class ResponseFactory implements ResponseFactoryInterface
 {
+    public function __construct()
+    {
+    }
+
     public function createResponse(int $code = 200, string $reasonPhrase = ''): ResponseInterface
     {
         return new Response(