+++ tests/LoadBrickTest.php
@@ -31,6 +31,8 @@ use Marmotte\Brick\Bricks\BrickLoader;
 use Marmotte\Brick\Bricks\BrickManager;
 use Marmotte\Brick\Cache\CacheManager;
 use Marmotte\Brick\Mode;
+use Marmotte\Http\Request\ServerRequest;
+use Marmotte\Http\Response\ResponseFactory;
 use PHPUnit\Framework\TestCase;
 
 class LoadBrickTest extends TestCase
@@ -43,10 +45,36 @@ class LoadBrickTest extends TestCase
             new CacheManager(mode: Mode::TEST)
         );
         $brick_loader->loadFromDir(__DIR__ . '/../src');
+        $service_manager = $brick_manager->initialize(__DIR__ . '/../src', __DIR__ . '/../src');
 
         $bricks = $brick_manager->getBricks();
         self::assertCount(1, $bricks);
         $brick = $bricks[0];
         self::assertSame(HttpBrick::class, $brick->brick->getName());
+
+        self::assertTrue($service_manager->hasService(ResponseFactory::class));
+        self::assertTrue($service_manager->hasService(ServerRequest::class));
+    }
+
+    public function testItLoadUriForServerRequest(): void
+    {
+        $_SERVER = [
+            'HTTPS'       => 'on',
+            'HTTP_HOST'   => 'example.com',
+            'REQUEST_URI' => '/path/to/page?arg=value#anchor',
+        ];
+
+        $brick_manager = new BrickManager();
+        $brick_loader  = new BrickLoader(
+            $brick_manager,
+            new CacheManager(mode: Mode::TEST)
+        );
+        $brick_loader->loadFromDir(__DIR__ . '/../src');
+        $service_manager = $brick_manager->initialize(__DIR__ . '/../src', __DIR__ . '/../src');
+
+        self::assertTrue($service_manager->hasService(ServerRequest::class));
+        $request = $service_manager->getService(ServerRequest::class);
+        self::assertInstanceOf(ServerRequest::class, $request);
+        self::assertSame('https://example.com/path/to/page?arg=value#anchor', (string) $request->getUri());
     }
 }